requirements.txt
Django==2.1.1
djangorestframework==3.8.2
PuLP==1.6.8